pyproject.toml
[project]
name = "helix-hive"
version = "0.1.0"
description = "Multi-agent system for collaborative protein engineering"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython",
    "torch",                # needed for esm/protgpt2
    "transformers",
    "fair-esm",                  # Meta’s ESM models - Note `esm` is the modern lib but doesn't have esm2 models, hence installing the older dist
    "langchain",
    "langgraph",
    "streamlit",
    "chroma-hnswlib",       # local vector DB
    "requests",
    "numpy",
    "matplotlib",
]

[dependency-groups]  # uv-specific config
dev = [
    "ruff",
    "pytest",
    "ipython",
]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "B"]        # pyflakes, pycodestyle, bugbear
ignore = ["E501"]               # optional: ignore long lines